# contracts/sentinel/v1.8/openapi.json
{
  "components": {
    "schemas": {
...
      "CommandRequest": {
        "additionalProperties": false,
        "properties": {
          "commandId": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
            "title": "Commandid",
            "type": "string"
          },
          "holderId": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
            "title": "Holderid",
            "type": "string"
          },
          "intent": {
            "$ref": "#/components/schemas/Intent"
          }
        },
        "required": [
          "commandId",
          "holderId",
          "intent"
        ],
        "title": "CommandRequest",
        "type": "object"
      },
...
      "CreateRunRequest": {
        "additionalProperties": false,
        "properties": {
          "creationId": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
            "title": "Creationid",
            "type": "string"
          },
          "scenario": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScenarioRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "templateId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Templateid"
          }
        },
        "required": [
          "creationId"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LegacyD2Receipt": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "code": {
            "enum": [
              "OK",
              "NOT_INTERACTIVE",
              "DEMO_DISABLED",
              "ACTIVE_RUN_EXISTS",
              "IDENTITY_CONFLICT",
              "INTENT_INVALID",
              "INTENT_EXPIRED",
              "OBSOLETE_INTENT",
              "REFERENCE_MISMATCH",
              "CONTROL_HELD",
              "CONTROL_REQUIRED",
              "LEASE_EXPIRED",
              "RECLAIM_REQUIRED",
              "INVALID_TRANSITION",
              "RUN_TERMINAL",
              "INVALID_REQUEST",
              "NOT_FOUND",
              "SOURCE_UNHEALTHY",
              "FRAME_INVALID",
              "MOVE_EXPIRED",
              "SELECTION_INVALID",
              "BINDING_CHANGED",
              "ASSET_BUSY",
              "POSITION_UNAVAILABLE",
              "UNSUPPORTED_REFERENCE",
              "OUTSIDE_EXTENT",
              "ENDPOINT_INVALID",
              "EXECUTION_TERMINAL",
              "NO_AVAILABLE_ASSETS",
              "ORDER_SUPERSEDED"
            ],
            "title": "Code",
            "type": "string"
          },
          "directOrder": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directorder"
          },
          "executionIds": {
            "items": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
              "type": "string"
            },
            "maxItems": 32,
            "title": "Executionids",
            "type": "array"
          },
          "frameId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frameid"
          },
          "memberOutcomes": {
            "items": {
              "$ref": "#/components/schemas/LegacyD2DirectMemberOutcome"
            },
            "maxItems": 32,
            "title": "Memberoutcomes",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "missionId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missionid"
          },
          "operation": {
            "enum": [
              "create",
              "acquire",
              "renew",
              "reclaim",
              "revoke",
              "start",
              "pause",
              "resume",
              "end",
              "move",
              "cancel",
              "direct-move"
            ],
            "title": "Operation",
            "type": "string"
          },
          "recordedAt": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
            "title": "Recordedat",
            "type": "string"
          },
          "recordingId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recordingid"
          },
          "requestId": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
            "title": "Requestid",
            "type": "string"
          },
          "runId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runid"
          },
          "schemaVersion": {
            "const": "1.2",
            "default": "1.2",
            "title": "Schemaversion",
            "type": "string"
          },
          "sequence": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          }
        },
        "required": [
          "requestId",
          "operation",
          "accepted",
          "code",
          "message",
          "recordedAt"
        ],
        "title": "LegacyD2Receipt",
        "type": "object"
      },
...
      "Receipt": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "code": {
            "enum": [
              "OK",
              "NOT_INTERACTIVE",
              "DEMO_DISABLED",
              "ACTIVE_RUN_EXISTS",
              "IDENTITY_CONFLICT",
              "INTENT_INVALID",
              "INTENT_EXPIRED",
              "OBSOLETE_INTENT",
              "REFERENCE_MISMATCH",
              "CONTROL_HELD",
              "CONTROL_REQUIRED",
              "LEASE_EXPIRED",
              "RECLAIM_REQUIRED",
              "INVALID_TRANSITION",
              "RUN_TERMINAL",
              "INVALID_REQUEST",
              "NOT_FOUND",
              "SOURCE_UNHEALTHY",
              "FRAME_INVALID",
              "MOVE_EXPIRED",
              "SELECTION_INVALID",
              "BINDING_CHANGED",
              "ASSET_BUSY",
              "POSITION_UNAVAILABLE",
              "UNSUPPORTED_REFERENCE",
              "OUTSIDE_EXTENT",
              "ENDPOINT_INVALID",
              "EXECUTION_TERMINAL",
              "NO_AVAILABLE_ASSETS",
              "ORDER_SUPERSEDED"
            ],
            "title": "Code",
            "type": "string"
          },
          "controlOrder": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Controlorder"
          },
          "controlOutcomes": {
            "items": {
              "$ref": "#/components/schemas/ControlMemberOutcome"
            },
            "maxItems": 32,
            "title": "Controloutcomes",
            "type": "array"
          },
          "directOrder": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Directorder"
          },
          "executionIds": {
            "items": {
              "maxLength": 256,
              "minLength": 1,
              "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
              "type": "string"
            },
            "maxItems": 32,
            "title": "Executionids",
            "type": "array"
          },
          "frameId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frameid"
          },
          "memberOutcomes": {
            "items": {
              "$ref": "#/components/schemas/DirectMemberOutcome"
            },
            "maxItems": 32,
            "title": "Memberoutcomes",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "missionId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Missionid"
          },
          "operation": {
            "enum": [
              "create",
              "acquire",
              "renew",
              "reclaim",
              "revoke",
              "start",
              "pause",
              "resume",
              "end",
              "move",
              "cancel",
              "direct-move",
              "stop",
              "return-to-script"
            ],
            "title": "Operation",
            "type": "string"
          },
          "recordedAt": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{3}Z$",
            "title": "Recordedat",
            "type": "string"
          },
          "recordingId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recordingid"
          },
          "requestId": {
            "maxLength": 256,
            "minLength": 1,
            "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
            "title": "Requestid",
            "type": "string"
          },
          "runId": {
            "anyOf": [
              {
                "maxLength": 256,
                "minLength": 1,
                "pattern": "^[^\\s\\x00-\\x1f\\x7f](?:[^\\x00-\\x1f\\x7f]*[^\\s\\x00-\\x1f\\x7f])?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runid"
          },
          "schemaVersion": {
            "const": "1.3",
            "default": "1.3",
            "title": "Schemaversion",
            "type": "string"
          },
          "sequence": {
            "anyOf": [
              {
                "maximum": 9007199254740991.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sequence"
          }
        },
        "required": [
          "requestId",
          "operation",
          "accepted",
          "code",
          "message",
          "recordedAt"
        ],
        "title": "Receipt",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Sentinel world authority",
    "version": "1.7.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/interactive/creations": {
      "get": {
        "operationId": "creation_receipt_api_interactive_creations_get",
        "parameters": [
          {
            "in": "query",
            "name": "identity",
            "required": true,
            "schema": {
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyD2Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyM12Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyReceipt"
                    }
                  ],
                  "title": "Response Creation Receipt Api Interactive Creations Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Creation Receipt",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/creations/{identity}": {
      "get": {
        "deprecated": true,
        "operationId": "creation_receipt_api_interactive_creations__identity__get",
        "parameters": [
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyD2Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyM12Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyReceipt"
                    }
                  ],
                  "title": "Response Creation Receipt Api Interactive Creations  Identity  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Creation Receipt",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
...
    "/api/interactive/runs": {
      "post": {
        "operationId": "create_api_interactive_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyD2Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyM12Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyReceipt"
                    }
                  ],
                  "title": "Response Create Api Interactive Runs Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/{mission_id}/commands": {
      "post": {
        "operationId": "command_api_interactive__mission_id__commands_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyD2Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyM12Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyReceipt"
                    }
                  ],
                  "title": "Response Command Api Interactive  Mission Id  Commands Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Command",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/{mission_id}/direct-moves": {
      "post": {
        "operationId": "direct_move_api_interactive__mission_id__direct_moves_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DirectMoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyD2Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyM12Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyReceipt"
                    }
                  ],
                  "title": "Response Direct Move Api Interactive  Mission Id  Direct Moves Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Direct Move",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/{mission_id}/executions": {
      "get": {
        "operationId": "executions_api_interactive__mission_id__executions_get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "throughFrameId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Throughframeid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExecutionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Executions",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/{mission_id}/intents": {
      "post": {
        "operationId": "intent_api_interactive__mission_id__intents_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Intent"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Intent",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/{mission_id}/moves": {
      "post": {
        "operationId": "move_api_interactive__mission_id__moves_post",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyD2Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyM12Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyReceipt"
                    }
                  ],
                  "title": "Response Move Api Interactive  Mission Id  Moves Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Move",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/{mission_id}/receipts": {
      "get": {
        "operationId": "receipt_api_interactive__mission_id__receipts_get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "identity",
            "required": true,
            "schema": {
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyD2Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyM12Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyReceipt"
                    }
                  ],
                  "title": "Response Receipt Api Interactive  Mission Id  Receipts Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receipt",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/{mission_id}/receipts/{identity}": {
      "get": {
        "deprecated": true,
        "operationId": "receipt_api_interactive__mission_id__receipts__identity__get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "identity",
            "required": true,
            "schema": {
              "title": "Identity",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyD2Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyM12Receipt"
                    },
                    {
                      "$ref": "#/components/schemas/LegacyReceipt"
                    }
                  ],
                  "title": "Response Receipt Api Interactive  Mission Id  Receipts  Identity  Get"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receipt",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
    "/api/interactive/{mission_id}/status": {
      "get": {
        "operationId": "status_api_interactive__mission_id__status_get",
        "parameters": [
          {
            "in": "path",
            "name": "mission_id",
            "required": true,
            "schema": {
              "title": "Mission Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Status",
        "tags": [
          "local synthetic demo"
        ]
      }
    },
...
  },
...
}